refactor(furl-core): move config and engine to core mod.
