+++ crates/furl-core/src/core/engine.rs
@@ -20,6 +20,8 @@ use tokio::{
     sync::Mutex,
 };
 
+use super::config::DownloadConfig;
+
 const _1MB: u64 = 1024 * 1024;
 const _10MB: u64 = 10 * 1024 * 1024;
 
@@ -52,29 +54,6 @@ impl Chunk {
             end_byte,
             downloaded: 0,
         }
-    }
-}
-
-#[derive(Clone)]
-pub struct DownloadConfig {
-    pub max_chunk_size: u64,
-}
-
-impl DownloadConfig {
-    pub fn new() -> Self {
-        Self {
-            max_chunk_size: _10MB,
-        }
-    }
-    pub fn set_max_chunk_size(mut self, size: u64) -> Self {
-        self.max_chunk_size = size;
-        self
-    }
-}
-
-impl Default for DownloadConfig {
-    fn default() -> Self {
-        Self::new()
     }
 }
